absorb _batch module to __init__
